fix(manage): finalize semantic authoring safeguards
